# docs/swagger.yaml
basePath: /
definitions:
...
  dto.ErrorResponse:
    properties:
      message:
        example: Please provide valid inputs
        type: string
      reason:
        example: detalhes do erro se houver
        type: string
    type: object
...
  dto.MakePaymentSuccess:
    properties:
      message:
        example: make payment
        type: string
      pubKey:
        example: exemplo_de_chave_privada_asd2asd2
        type: string
      secret:
        example: exemplo_de_segredo_stripe_asd2asd2
        type: string
    type: object
...
host: localhost:3333
info:
  contact:
    email: contato@otavioteixeira.com
    name: Otávio Augusto
    url: http://github.com/T4vexx
  description: API de e-commerce feita em Go com Fiber
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Ecommerce Go API
  version: "1.0"
paths:
  /buyer/payment:
    get:
      description: Creates a payment request for the authenticated buyer. If an active
        payment exists, returns its client secret; otherwise, creates a new payment.
      produces:
      - application/json
      responses:
        "200":
          description: Payment information including public key and client secret
          schema:
            $ref: '#/definitions/dto.MakePaymentSuccess'
        "400":
          description: Error generating payment or saving payment
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Error generating order id
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Initiates a payment
      tags:
      - Transaction
  /buyer/verify:
    get:
      description: Checks the status of an active payment for the authenticated buyer,
        updates the payment status, and creates an order if payment is successful.
      produces:
      - application/json
      responses:
        "200":
          description: Verification result along with payment response
          schema:
            additionalProperties: true
            type: object
        "400":
          description: No active payment exists or error during verification
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal server error during verification
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Verifies payment status
      tags:
      - Transaction
...
swagger: "2.0"